requirements.txt
aioredis==1.3.1
amqp==2.6.1
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
celery-progress==0.0.10
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
daphne==2.5.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-channels==0.7.0
django-composite-foreignkey==1.1.0
django-compositekey==1.5.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-timezone-field==4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
drf-yasg==1.17.1
filelock==3.0.12
flower==0.9.5
future==0.18.2
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
hiredis==1.1.0
httpie==2.2.0
humanize==2.5.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
pipenv==2020.6.2
prometheus-client==0.8.0
promise==2.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydevd-pycharm==202.6397.98
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Rx==1.6.1
service-identity==18.1.0
simplejson==3.17.2
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
stripe==2.55.0
toml==0.10.1
tornado==6.0.4
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.30
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.1.0
